Update _index.tsx
Major overhaul enabling searching for pokemon with spaces, periods, and colons in their name.

pokemonNameMap takes those queries and puts them in an api-friendly format when called by apiFriendlyName.

noralizeName replaces special characters with hyphens.

normalizedInput is the fallback name used to ensure the correct formatting is used for the api call.

the api call itself was updated to use the apiFriendlyName function.

Error logging improved.

Removed result2 and result2Json as they are unnecessary with the updated API logic.

Pokemon name filters were altered to include the query, rather than start with the query.

The Pokeball image was added as a fallback when no pokemon could be found.

handleKeyDown was created because while the updated API calls were being made correctly with apiFriendlyName, stats, types, and sprites were not being correctly rendered. This handler ensures that the API friendly name is used for all aspects of the GET request.

Slight misc text changes.