Update _index.tsx
Overhauled UI. Sprites and stats are now in conditional boxes that appear side by side, while the pokeball fallback image is centered. Sprite size needs to be modified to be larger but I imagine it's currently kept small because of a specificity issue or something similar. Stats will also need to be centered relative to the box. TODO: Style typing text to look like the games.